Handle aliased resource defintions.
When the RAML parser encounters an alias, it will use a reference to the
resulting json object everywhere the alias is encountered (if it can).
The console decorates the parser output which means that this reference
might be used in a way which would require it to be decorated in
conflicting ways.

For now, solving this by decorating a clone of the resource object from
the parser.